cucumber/pyproject.toml
[tool.poetry]
name = "cucumber"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
moto = "^5.0.13"
boto3 = "^1.35.0"
PyHamcrest = "^2.1.0"
behave = "^1.2.6"
bottle = "^0.12.25"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
